pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ha-boss"
version = "0.1.0"
description = "A Home Assistant management service for robust automation and management"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Jason Hagerty", email = "jasonthagerty@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "aiofiles>=23.2.0",
    "aiohttp>=3.9.0",
    "anthropic>=0.40.0,<1.0.0",
    "fastapi>=0.109.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "pyyaml>=6.0",
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.0",
    "flake8>=6.1.0",
    "ruff>=0.1.8",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "types-aiofiles>=23.2.0",
    "types-pyyaml>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/jasonthagerty/ha_boss"
Documentation = "https://github.com/jasonthagerty/ha_boss#readme"
Repository = "https://github.com/jasonthagerty/ha_boss"

[tool.setuptools]
packages = ["ha_boss"]

[project.scripts]
haboss = "ha_boss.cli.commands:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --strict-markers"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "performance: marks tests as performance benchmarks (deselect with '-m \"not performance\"')",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # function call in argument defaults (required for Typer)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
